requirements.txt
pyyaml
opencv-python
mediapipe
deepgram
requests
python-dotenv
watchdog
functools
transformers
tensorflow
tf-keras